+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.8.0-extra-common-dev
 
+- Support pre-populated database for Windows "Universal"
+
 ## 0.7.1-common-dev
 
 - Fix to Windows "Universal" version to support big integers